add orderservice and remove ordermanager
